refactor(middleware): replace `LivewireSentryContextMiddleware` with `LivewireContextMiddleware`
- Remove `LivewireSentryContextMiddleware` class and its associated logic.
- Add new `LivewireContextMiddleware` with simplified context handling using `Illuminate\Support\Facades\Context`.
- Update `LivewireServiceProvider` to register `LivewireContextMiddleware` for Livewire update routes.
